refactor: remove all internal usage of `executeCommand` helper
This is a relatively large refactor that involves the following
changes:

- remove ExecuteDbAdminCommandOperation
- remove the remaining legacy index operations, convert them into
  `CreateIndex`/`CreateIndexes`/`EnsureIndex` operations, and put
  all index-related operations into a single file
- convert find_and_modify to command operation, also merge all
  FindOneAnd* operations into a single file
- remove all internal usage of `executeCommand` helper and the
  helper itself

Note that a copy of `executeCommand` is perserved in the
FindAndModifyOperation because a larger refactor is needed to port
write commands to the operations layer

NODE-2714
