+++ src/operations/count.ts
@@ -1,6 +1,6 @@
 import { Aspect, defineAspects } from './operation';
-import { buildCountCommand } from './common_functions';
 import CommandOperation = require('./command');
+import { decorateWithCollation, decorateWithReadConcern } from '../utils';
 
 class CountOperation extends CommandOperation {
   cursor: any;
@@ -51,6 +51,51 @@ class CountOperation extends CommandOperation {
       callback(err, result ? result.n : null);
     });
   }
+}
+
+/**
+ * Build the count command.
+ *
+ * @function
+ * @param {Collection|Cursor} collectionOrCursor an instance of a collection or cursor
+ * @param {any} query The query for the count.
+ * @param {any} [options] Optional settings. See Collection.prototype.count and Cursor.prototype.count for a list of options.
+ */
+function buildCountCommand(collectionOrCursor: any, query: any, options?: any) {
+  const skip = options.skip;
+  const limit = options.limit;
+  let hint = options.hint;
+  const maxTimeMS = options.maxTimeMS;
+  query = query || {};
+
+  // Final query
+  const cmd = {
+    count: options.collectionName,
+    query: query
+  } as any;
+
+  if (collectionOrCursor.s.numberOfRetries) {
+    // collectionOrCursor is a cursor
+    if (collectionOrCursor.options.hint) {
+      hint = collectionOrCursor.options.hint;
+    } else if (collectionOrCursor.cmd.hint) {
+      hint = collectionOrCursor.cmd.hint;
+    }
+    decorateWithCollation(cmd, collectionOrCursor, collectionOrCursor.cmd);
+  } else {
+    decorateWithCollation(cmd, collectionOrCursor, options);
+  }
+
+  // Add limit, skip and maxTimeMS if defined
+  if (typeof skip === 'number') cmd.skip = skip;
+  if (typeof limit === 'number') cmd.limit = limit;
+  if (typeof maxTimeMS === 'number') cmd.maxTimeMS = maxTimeMS;
+  if (hint) cmd.hint = hint;
+
+  // Do we have a readConcern specified
+  decorateWithReadConcern(cmd, collectionOrCursor);
+
+  return cmd;
 }
 
 defineAspects(CountOperation, [
